\name{frameMe}
\alias{frameMe}

\title{Extracting column "Median" from all data frames in a list.}

\description{This function assembles together the plot ready target-pathway specific values from all diseases, for all samples, for each tool, into a single data frame.}

\usage{frameMe(x)}

\arguments{\item{x}{list of resultant data frames}}

\value{\item{df}{A dataframe of median values}}

\author{Shaurya Jauhari}
