\name{installBenchmarkPackages}
\alias{installBenchmarkPackages}


\title{Installing requisite packages for the benchmarking protocol}

\description{The function will install the package dependencies for benchmarking tools, viz. Seq2pathway, Chipenrich and Broadenrich and set up the work environment.}

\usage{installBenchmarkPackages()}

\author{Shaurya Jauhari}

\examples{installBenchmarkPackages()}

