\name{plotMetrics}
\alias{plotMetrics}

\title{Generate ggplots for the given metric}

\description{The function is used to plot the median values returned by the "frameMe" function. The ggplot is a bar plot for the particular metric against all tools in question, i.e. Broadenrich, Chipenrich, Enrichr, GREAT, and Seq2pathway. 
}

\usage{plotMetrics(x)}

\arguments{\item{x}{A dataframe with median values for all samples(for all diseases), under each tool.}}

\author{Shaurya Jauhari}
